+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### 🛠 Improvements
 - [5318](https://github.com/vegaprotocol/vega/issues/5318) - Automatically dispatch reward pool into markets in recurring transfers
 - [5333](https://github.com/vegaprotocol/vega/issues/5333) - Run snapshot generation for all providers in parallel 
+- [5324](https://github.com/vegaprotocol/vega/issues/5324) -  Send event when oracle data doesn't match
 
 ### 🐛 Fixes
 - [5277](https://github.com/vegaprotocol/vega/issues/5277) - Updating a market should default auction extension to 1